se creó el hook y se pasó todas las funciones al hook para tener el codigo más limpio
