+++ src/App.jsx
@@ -1,82 +1,22 @@
-import { useState, useEffect } from "react";
-import { db } from "./database/db.js";
 import Footer from "./components/Footer";
 import Guitar from "./components/Guitar";
 import Header from "./components/Header";
+import { useCart } from "./hooks/useCart.js";
 
 function App() {
 
-  const initialCart = () => {
-    const localStorageCart = localStorage.getItem('cart')
-    return localStorageCart ? JSON.parse(localStorageCart) : []
-  }
-
-  const [data] = useState(db);
-  const [cart, setCart] = useState(initialCart);
-
-  const MAX_ITEMS = 5
-  const MIN_ITEMS = 1
-
-  useEffect(() => {
-    localStorage.setItem('cart', JSON.stringify(cart))
-  }, [cart])
-
-  function addToCart(item) {
-    const itemExists = cart.findIndex((guitar) => guitar.id === item.id);
-    if (itemExists >= 0) {
-      if(cart[itemExists].quantity >= MAX_ITEMS) return
-      const updateCart = [...cart];
-      updateCart[itemExists].quantity++;
-      setCart(updateCart);
-    } else {
-      item.quantity = 1;
-      setCart([...cart, item]);
-    }
-    
-  }
-
-  function removeFromCart(id) {
-    setCart((prevCart) => prevCart.filter(guitar => guitar.id !== id))
-  }
-
-  function increaseQuantity(id) {
-    const updateCart = cart.map( item => {
-      if(item.id === id && item.quantity < MAX_ITEMS) {
-        return {
-          ...item,
-          quantity: item.quantity + 1
-        }
-      }
-      return item
-    })
-    setCart(updateCart)
-  }
-
-  function decreaseQuantity(id) {
-    const updateCart = cart.map( item => {
-      if(item.id === id && item.quantity > MIN_ITEMS) {
-        return {
-          ...item,
-          quantity: item.quantity - 1
-        }
-      }
-      return item
-    })
-    setCart(updateCart)
-  }
-
-  function clearCart() {
-    setCart([])
-  }
+  const { cart, removeFromCart, increaseQuantity, decreaseQuantity, clearCart, data, addToCart, isEmpty, cartTotal } = useCart()
 
   return (
     <>
-      <Header 
+      <Header
         cart={cart}
         removeFromCart={removeFromCart}
         increaseQuantity={increaseQuantity}
         decreaseQuantity={decreaseQuantity}
         clearCart={clearCart}
+        isEmpty={isEmpty}
+        cartTotal={cartTotal}
       />
 
       <main className="container-xl mt-5">
@@ -87,7 +27,6 @@ function App() {
             <Guitar
               key={guitar.id}
               guitar={guitar}
-              setCart={setCart}
               addToCart={addToCart}
             />
           ))}
